Cast sprintf args in sink.c to expected types (to avoid problems on systems where pid_t is not int) (John Bley <jbb6@acpub.duke.edu>)
svn path=/trunk/; revision=2465

+++ sink.c
@@ -282,7 +282,7 @@ static int send_bouncemail(struct query *ctl, struct msgblk *msg,
 
     sprintf(boundary, 
 	    "om-mani-padme-hum-%d-%d-%ld", 
-	    getpid(), getppid(), time((time_t *)NULL));
+	    (int)getpid(), (int)getppid(), time((time_t *)NULL));
 
     ts = rfc822timestamp();
